* commit '4dc6545fa40fea99301a048f5350cca486c64c46':
  YUK-24499 : Add validations on loggers(fixed review comment)
  YUK-24499 : Add validations on loggers(fixed date issue and system logger delete and refactoring)
  YUK-24499 : Add validations on loggers(fixed issues found in verification and other issues)

+++ yukon-client/common/src/com/cannontech/clientutils/logger/service/impl/YukonLoggerServiceImpl.java
@@ -10,7 +10,10 @@ import com.cannontech.clientutils.logger.dao.YukonLoggerDao;
 import com.cannontech.clientutils.logger.service.YukonLoggerService;
 import com.cannontech.common.api.token.ApiRequestContext;
 import com.cannontech.common.events.loggers.SystemEventLogService;
+import com.cannontech.common.exception.DeletionFailureException;
 import com.cannontech.common.log.model.LoggerLevel;
+import com.cannontech.common.log.model.LoggerType;
+import com.cannontech.common.log.model.SystemLogger;
 import com.cannontech.common.log.model.YukonLogger;
 import com.cannontech.common.model.Direction;
 import com.cannontech.core.dao.NotFoundException;
@@ -26,7 +29,9 @@ public class YukonLoggerServiceImpl implements YukonLoggerService {
     @Override
     public YukonLogger getLogger(int loggerId) {
         try {
-            return yukonLoggerDao.getLogger(loggerId);
+            YukonLogger logger = yukonLoggerDao.getLogger(loggerId);
+            logger.setLoggerType(SystemLogger.isSystemLogger(logger.getLoggerName()) ? LoggerType.SYSTEM_LOGGER : LoggerType.USER_LOGGER );
+            return logger;
         } catch (EmptyResultDataAccessException e) {
             throw new NotFoundException("Logger Id not found");
         }
@@ -55,6 +60,9 @@ public class YukonLoggerServiceImpl implements YukonLoggerService {
     public int deleteLogger(int loggerId) {
         try {
             String loggerName = getLogger(loggerId).getLoggerName();
+            if(SystemLogger.isSystemLogger(loggerName)) {
+                throw new DeletionFailureException("System logger deletion not supported.");
+            }
             yukonLoggerDao.deleteLogger(loggerId);
             dbChangeManager.processDbChange(DbChangeType.DELETE, DbChangeCategory.LOGGER, loggerId);
             systemEventLogService.loggerDeleted(loggerName, ApiRequestContext.getContext().getLiteYukonUser());
